+++ components/MapLogic/MapLogic.js
@@ -1,10 +1,11 @@
 import Map from "../Map/Map";
-import "mapbox-gl/dist/mapbox-gl.css";
 import postData from "@/lib/postData";
+import { useState, useEffect } from "react";
 import createDbData from "@/lib/createDbData";
 import drawRouteOnMap from "@/lib/drawRouteOnMap";
 import coordsToString from "@/lib/coordsToString";
-import { useRef, useState, useEffect } from "react";
+import { useMapboxMap } from "@/hooks/useMapboxMap";
+import { useMapMarkers } from "@/hooks/useMapMarkers";
 import { useGeocodingApi } from "@/hooks/useGeocodingApi";
 import { useDirectionsApi } from "@/hooks/useDirectionsApi";
 import mapboxgl from "!mapbox-gl"; // eslint-disable-line import/no-webpack-loader-syntax
@@ -13,59 +14,31 @@ import mapboxgl from "!mapbox-gl"; // eslint-disable-line import/no-webpack-load
 const accessToken =
   "pk.eyJ1Ijoia2V0bmVrIiwiYSI6ImNsaDY5MG1rdjAzdGQzZW5zanhqbWU5cmoifQ.jOgkeSYbOWlO7yvv_sA0Dg";
 
-export default function MapHandler() {
-  const map = useRef(null);
-  const mapContainer = useRef(null);
-  const [markers, setMarkers] = useState([]);
+export default function MapLogic() {
+  const [map, mapContainer] = useMapboxMap();
   const [searchValue, setSearchValue] = useState("");
-  const [routeCoords, setRouteCoords] = useState([]);
   const [calculated, setCalculated] = useState(false);
   const [savedRoute, setSavedRoute] = useState(false);
   const { searchResults, getGeocodingData } = useGeocodingApi();
   const [inputPlaceholder, setInputPlaceholder] = useState("Search for places");
   const { isLoading, directionsData, setDirectionsData, getDirectionsData } =
     useDirectionsApi();
+  const { markers, routeCoords, setMarkers, setRouteCoords } =
+    useMapMarkers(map);
 
   mapboxgl.accessToken = accessToken;
 
-  //Map and Markers
   useEffect(() => {
-    map.current = new mapboxgl.Map({
-      container: mapContainer.current,
-      style: "mapbox://styles/mapbox/streets-v12",
-      center: [10.000654, 53.550341],
-      zoom: 9,
-    });
-
-    map.current.on("click", function (event) {
-      const markerCoords = [event.lngLat.lng, event.lngLat.lat];
-      setRouteCoords((prevCoords) => [...prevCoords, markerCoords]);
-      const marker = new mapboxgl.Marker()
-        .setLngLat(markerCoords)
-        .addTo(map.current);
-
-      setMarkers((prevMarkers) => [...prevMarkers, marker]);
-    });
-  }, []);
-
-  useEffect(() => {
+    // Only draw when directions data available
     if (!isLoading && directionsData !== null) {
       // Get the coordinates from the response
       const coords = directionsData.routes[0].geometry;
       // Draw the route on the map
       drawRouteOnMap(coords, map);
     }
-  }, [isLoading, directionsData]);
+  }, [isLoading, directionsData, map]);
 
   // all handler
-
-  function handleCreateClick() {
-    const coordsString = coordsToString(routeCoords);
-    getDirectionsData(
-      `https://api.mapbox.com/directions/v5/mapbox/cycling/${coordsString}?geometries=geojson&language=de&overview=full&steps=true&access_token=${mapboxgl.accessToken}`
-    );
-    setCalculated(true);
-  }
 
   function handleSaveClick() {
     setSavedRoute(true);
@@ -117,6 +90,12 @@ export default function MapHandler() {
     setCalculated(false);
   }
 
+  function handleResultClick(resultCoords, placeholderText) {
+    map.current.flyTo({ center: resultCoords, zoom: 12 });
+    setInputPlaceholder(placeholderText);
+    setSearchValue("");
+  }
+
   function handleSearchSubmit(event) {
     event.preventDefault();
     if (searchValue === "") return;
@@ -126,10 +105,12 @@ export default function MapHandler() {
     setInputPlaceholder(searchResults[0]?.place_name);
   }
 
-  function handleResultClick(resultCoords, placeholderText) {
-    map.current.flyTo({ center: resultCoords, zoom: 12 });
-    setInputPlaceholder(placeholderText);
-    setSearchValue("");
+  function handleCreateClick() {
+    const coordsString = coordsToString(routeCoords);
+    getDirectionsData(
+      `https://api.mapbox.com/directions/v5/mapbox/cycling/${coordsString}?geometries=geojson&language=de&overview=full&steps=true&access_token=${mapboxgl.accessToken}`
+    );
+    setCalculated(true);
   }
 
   function handleInputChange(event) {
@@ -143,20 +124,19 @@ export default function MapHandler() {
       markers={markers}
       isLoading={isLoading}
       calculated={calculated}
-      map={map}
+      savedRoute={savedRoute}
+      searchValue={searchValue}
       mapContainer={mapContainer}
+      searchResults={searchResults}
+      inputPlaceholder={inputPlaceholder}
       onSave={handleSaveClick}
       onCreate={handleCreateClick}
       onDelete={handleDeleteClick}
-      searchValue={searchValue}
-      searchResults={searchResults}
-      inputPlaceholder={inputPlaceholder}
       onInputChange={handleInputChange}
       onClickResult={handleResultClick}
-      onSearchSubmit={handleSearchSubmit}
-      savedRoute={savedRoute}
       onCancelClick={handleCancelClick}
       onRouteSubmit={handleRouteSubmit}
+      onSearchSubmit={handleSearchSubmit}
     />
   );
 }
