+++ lib/postData.js
@@ -9,11 +9,9 @@ export default async function postData(dbData) {
     });
 
     if (!response.ok) {
-      // Failure (Bad Response)
       console.error("Bad Response");
     }
   } catch (error) {
-    // Failure (Network error, etc)
     console.error("An Error occurred");
   }
 }
