--- lib/test.js
@@ -1,20 +1,0 @@
-createDbData();
-postData(dbData);
-
-//-----------------------//
-routeCoords;
-
-coordsToString / updateRoute;
-getMatch;
-
-//------------------------//
-
-//state
-isLoading;
-newRouteData;
-
-//calls
-getMatch(url);
-addRoute(coords);
-
-//--------------------//
